add: LayoutTypes
