+++ components/organisms/layout.tsx
@@ -1,7 +1,11 @@
 import Header from '@/components/organisms/header';
 import Footer from '@/components/organisms/footer';
 
-export default function Layout({ children }: any) {
+interface LayoutProps {
+  children: React.ReactNode;
+}
+
+export default function Layout({ children }: LayoutProps) {
   return (
     <>
       <div className="sticky top-0 z-40">
