--- lib/types.ts
@@ -1,11 +1,0 @@
-export interface BreadcrumbPage {
-  name: string;
-  href: string;
-  current: boolean;
-}
-
-export interface Sns {
-  name: string;
-  href: string;
-  icon: any;
-}
